requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autocorrect==2.6.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
build==1.2.1
CacheControl==0.14.0
cachetools==5.4.0
camelot-py==0.11.0
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
cleo==2.1.0
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
crashtest==0.4.1
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
datasets==2.21.0
datefinder==0.7.3
dateparser==1.2.0
DAWG-Python==0.7.2
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
dulwich==0.21.7
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.112.2
fastjsonschema==2.20.0
filelock==3.15.1
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
ghostscript==0.7
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.2.1
installer==0.7.0
intel-openmp==2021.4.0
intervaltree==3.1.0
ipykernel==6.29.4
ipymarkup==0.9.0
ipython==8.26.0
isodate==0.6.1
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
narwhals==1.6.2
natasha==1.6.0
navec==0.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.4
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.2
pefile==2023.2.7
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.2
prov==2.0.0
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyaspeller==2.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydot==2.0.0
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
pyproject_hooks==1.1.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-crf==0.7.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
razdel==0.5.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0-py3-none-any.whl#sha256=4b1c291c9b8991b708dadbf1b59f0fdaf7e70ca25e94df7fd3d523c4d5049754
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.0.1
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
slovnet==0.6.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.4
streamlit==1.37.0
sympy==1.12.1
tabula-py==2.9.3
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.5.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.42.3
trove-classifiers==2024.5.22
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.30.6
virtualenv==20.26.3
wasabi==1.1.3
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
xlrd==2.0.1
xxhash==3.5.0
yargy==0.16.0
yarl==1.9.11
zipp==3.19.2
frontend
pymupdf
mapply==0.1.23
PyMuPDF==1.24.5
PyPDF2==3.0.1
fuzzywuzzy==0.18.0